update dlp references in cargo
