+++ rust/Cargo.toml
@@ -33,8 +33,7 @@ ephemeral-rollups-sdk-attribute-action = { path = "action-attribute", version = 
 
 
 # Magicblock
-# magicblock-delegation-program-api = { version = "2.0.0", default-features = false }
-magicblock-delegation-program-api = { path = "/Users/snawaz/projects/mb/delegation-program/dlp-api", default-features = false }
+magicblock-delegation-program-api = { version = "3.0.0", default-features = false }
 magicblock-magic-program-api = { version = "0.8.5", default-features = false }
 
 ## External crates
